Moved all report-related functionality to new report form.
