+++ src/app/components/popUp/popUp.component.html
@@ -17,6 +17,13 @@
 			(editMode)="exitEdit()">
 		</display-name-edit-form>
 
+		<report-form
+			*ngIf="report"
+			[reportedItem]="reportedItem"
+			[reportType]="reportType"
+			(reportMode)="exitEdit()">
+		</report-form>
+
 		<!-- If the user is deleting something -->
 		<div *ngIf="delete && toDelete != 'ad post'" id="deleteItem" class="deleteItem">
 			<div class="popupText">Are you sure you want to delete this item?</div>
@@ -38,71 +45,6 @@
 				<button class="appButton deleteButton popupDeleteBtn" (click)="deletePost(false)">Delete without Closing Report</button>
 				<button class="appButton popupDeleteBtn" (click)="exitEdit()">Never Mind</button>
 			</div>
-		</div>
-
-		<!-- If the user is reporting a post -->
-		<div *ngIf="report && reportType == 'Post'" id="reportItem">
-			<h3 class="popupTitle">Report Post</h3>
-			<div class="userPost">
-				<div class="itemText" id="reportText">{{ reportedItem.text }}</div>
-				<div class="itemUser">By {{ reportedItem.user }}</div>
-			</div>
-			<div id="reportQ" class="popupText required">What is the problem with this post?</div>
-			<form id="reportPost" role="radiogroup">
-				<div class="formElement">
-					<input type="radio" class="radioButton" name="postReport" id="pRadioOption0" value="0" (click)="setSelected($event.target.value)" aria-describedby="reportQ">
-					<label class="radioButtonLabel" for="pRadioOption0">It is inppopriate.</label>
-				</div>
-				<div class="formElement">
-					<input type="radio" class="radioButton" name="postReport" id="pRadioOption1" value="1" (click)="setSelected($event.target.value)" aria-describedby="reportQ">
-					<label class="radioButtonLabel" for="pRadioOption1">It is spam.</label>
-				</div>
-				<div class="formElement">
-					<input type="radio" class="radioButton" name="postReport" id="pRadioOption2" value="2" (click)="setSelected($event.target.value)" aria-describedby="reportQ">
-					<label class="radioButtonLabel" for="pRadioOption2">It is offensive.</label>
-				</div>
-				<div class="formElement">
-					<fieldset>
-						<legend class="hidden" aria-hidden="false" id="rPostLegend">There's another problem with this post:</legend>
-						<input type="radio" class="radioButton" name="postReport" id="pRadioOption3" value="3" (click)="setSelected($event.target.value)" aria-describedby="reportQ">
-						<label class="radioButtonLabel" for="pRadioOption3">Other:</label> <input type="text" placeholder="Enter the response" id="rOption3Text" aria-labelledby="rPostLegend" disabled minlength="0" maxlength="120" aria-required="false">
-					</fieldset>
-				</div>
-
-				<div id="reportPText" class="labels">{{ 'report post ' + reportedItem.text }}</div>
-				<button type="submit" class="appButton reportButton" (click)="reportPost($event)" aria-labelledby="reportPText">Report Post</button>
-			</form>
-		</div>
-
-		<!-- If the user is reporting a post -->
-		<div *ngIf="report && reportType == 'User'" id="reportUser">
-			<h3 class="popupTitle">Report User</h3>
-			<a [routerLink]="['/user', reportedItem.userId]" id="uReportText">{{ reportedItem.displayName }}</a>
-			<div id="reportUQ" class="popupText required">What is the problem with this user?</div>
-			<form id="reportUserForm">
-				<div class="formElement">
-					<input type="radio" class="radioButton" name="postReport" id="uRadioOption0" value="0" (click)="setSelected($event.target.value)" aria-describedby="reportUQ">
-					<label class="radioButtonLabel" for="uRadioOption0">They are posting spam.</label>
-				</div>
-				<div class="formElement">
-					<input type="radio" class="radioButton" name="postReport" id="uRadioOption1" value="1" (click)="setSelected($event.target.value)" aria-describedby="reportUQ">
-					<label class="radioButtonLabel" for="uRadioOption1">They are posting harmful / dangerous content.</label>
-				</div>
-				<div class="formElement">
-					<input type="radio" class="radioButton" name="postReport" id="uRadioOption2" value="2" (click)="setSelected($event.target.value)" aria-describedby="reportUQ">
-					<label class="radioButtonLabel" for="uRadioOption2">They are behaving in an abusive manner.</label>
-				</div>
-				<div class="formElement">
-					<fieldset>
-						<legend class="hidden" aria-hidden="false" id="rUserLegend">There's another problem with this user:</legend>
-						<input type="radio" class="radioButton" name="postReport" id="uRadioOption3" value="3" (click)="setSelected($event.target.value)" aria-describedby="reportUQ">
-						<label class="radioButtonLabel" for="uRadioOption3">Other:</label> <input type="text" placeholder="Enter the response" id="uOption3Text" aria-labelledby="rUserLegend" disabled minlength="0" maxlength="120" aria-required="false">
-					</fieldset>
-				</div>
-
-				<div id="reportUText" class="labels">{{ 'Report user ' + reportedItem.displayName }}</div>
-				<button type="submit" class="appButton reportButton" (click)="reportUser($event)" aria-labelledby="reportUText">Report User</button>
-			</form>
 		</div>
 	</div>
 </div>
